requirements.txt
gradio>=3.40.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
reportlab>=3.6.0